EM's changes
I'm not sure if there's a misconception about this piece. I was told that this is how to get started using software in RHM and that FP+ is the example that others can follow. Is that not he case? I rewrote the introduction that way, so just let me know. 

Other changes:
Make it clear from the top that RHM is for RH OS 4 users
Led with the problem: Getting started with the trial software isn’t easy

I’m very confused about the introduction of IBM Cloud in this piece. Do you have to have an IBM Cloud account to interact with RHM? If so, should that be in the prereqs. Or is this piece really about how to use IBM Cloud with RHM?  I’m afraid that the if this content is being written for people who use Red Hat OpenShift on IBM Cloud, then the audience might be too small. 

Now we are requiring readers to:
> Have an IBM Cloud account
> Have a Red Hat OpenSHift account
> Be using/migrated to OpenShift 4.3

Can we take IBM out of it?